Extracted extension from Certificate Manager
